requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
anyio==4.12.1
brotlicffi
certifi
cffi
charset-normalizer
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.1
filelock
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
gmpy2
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.6.0
idna
Jinja2
joblib==1.5.3
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.14
mkl-service==2.4.0
mkl_fft
mkl_random
mpmath
networkx
numpy
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
packaging
pillow
pycparser
Pygments==2.19.2
pyparsing==3.3.2
PySocks
python-dateutil==2.9.0.post0
PyYAML
requests
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sounddevice==0.5.5
sympy==1.13.1
threadpoolctl==3.6.0
timm==1.0.25
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.3
typer==0.24.1
typing_extensions
urllib3
win_inet_pton
